+++ packages/components/app-menu/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.52](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.4.51...@synerise/ds-app-menu@0.4.52) (2021-04-07)
+
+**Note:** Version bump only for package @synerise/ds-app-menu
+
+
+
+
+
 ## [0.4.51](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.4.50...@synerise/ds-app-menu@0.4.51) (2021-03-30)
 
 **Note:** Version bump only for package @synerise/ds-app-menu