+++ packages/components/avatar/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.13.2...@synerise/ds-avatar@0.13.3) (2021-04-14)
+
+
+### Bug Fixes
+
+* **avatar:** add auto tooltip typing ([50c3e15](https://github.com/synerise/synerise-design/commit/50c3e151352c3aec92fdd709fd8ecdf8cc53e851))
+
+
+
+
+
+## [0.13.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.13.1...@synerise/ds-avatar@0.13.2) (2021-04-13)
+
+
+### Bug Fixes
+
+* **avatar:** no hover effect if no tooltip or onClick ([46e5866](https://github.com/synerise/synerise-design/commit/46e58664a5b1d40af637621cfc8450718f807d41))
+
+
+
+
+
+## [0.13.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.13.0...@synerise/ds-avatar@0.13.1) (2021-04-07)
+
+**Note:** Version bump only for package @synerise/ds-avatar
+
+
+
+
+
 # [0.13.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.12.2...@synerise/ds-avatar@0.13.0) (2021-03-30)