+++ packages/components/description/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.38](https://github.com/Synerise/synerise-design/compare/@synerise/ds-description@0.2.37...@synerise/ds-description@0.2.38) (2021-04-14)
+
+**Note:** Version bump only for package @synerise/ds-description
+
+
+
+
+
+## [0.2.37](https://github.com/Synerise/synerise-design/compare/@synerise/ds-description@0.2.36...@synerise/ds-description@0.2.37) (2021-04-13)
+
+**Note:** Version bump only for package @synerise/ds-description
+
+
+
+
+
+## [0.2.36](https://github.com/Synerise/synerise-design/compare/@synerise/ds-description@0.2.35...@synerise/ds-description@0.2.36) (2021-04-07)
+
+**Note:** Version bump only for package @synerise/ds-description
+
+
+
+
+
 ## [0.2.35](https://github.com/Synerise/synerise-design/compare/@synerise/ds-description@0.2.34...@synerise/ds-description@0.2.35) (2021-03-30)
 
 **Note:** Version bump only for package @synerise/ds-description