+++ packages/components/operators/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.18](https://github.com/Synerise/synerise-design/compare/@synerise/ds-operators@0.3.17...@synerise/ds-operators@0.3.18) (2021-04-07)
+
+**Note:** Version bump only for package @synerise/ds-operators
+
+
+
+
+
+## [0.3.17](https://github.com/Synerise/synerise-design/compare/@synerise/ds-operators@0.3.16...@synerise/ds-operators@0.3.17) (2021-04-01)
+
+
+### Bug Fixes
+
+* **operators:** conditional chaining on activeGroup id ([c76349c](https://github.com/Synerise/synerise-design/commit/c76349c4d31541e40b1f58c7b20d5aec09311340))
+
+
+
+
+
 ## [0.3.16](https://github.com/Synerise/synerise-design/compare/@synerise/ds-operators@0.3.15...@synerise/ds-operators@0.3.16) (2021-03-30)
 
 **Note:** Version bump only for package @synerise/ds-operators