+++ packages/components/page-header/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.34](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.12.33...@synerise/ds-page-header@0.12.34) (2021-04-07)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
 ## [0.12.33](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.12.32...@synerise/ds-page-header@0.12.33) (2021-03-30)
 
 **Note:** Version bump only for package @synerise/ds-page-header