+++ packages/components/table/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.29](https://github.com/synerise/synerise-design/compare/@synerise/ds-table@0.33.28...@synerise/ds-table@0.33.29) (2021-04-14)
+
+**Note:** Version bump only for package @synerise/ds-table
+
+
+
+
+
+## [0.33.28](https://github.com/synerise/synerise-design/compare/@synerise/ds-table@0.33.27...@synerise/ds-table@0.33.28) (2021-04-13)
+
+**Note:** Version bump only for package @synerise/ds-table
+
+
+
+
+
+## [0.33.27](https://github.com/synerise/synerise-design/compare/@synerise/ds-table@0.33.26...@synerise/ds-table@0.33.27) (2021-04-07)
+
+
+### Bug Fixes
+
+* **table:** set sort menu width ([0e26464](https://github.com/synerise/synerise-design/commit/0e26464085a6c06d7d30fa8ea455b75e9acd4f52))
+
+
+
+
+
 ## [0.33.26](https://github.com/synerise/synerise-design/compare/@synerise/ds-table@0.33.25...@synerise/ds-table@0.33.26) (2021-03-30)